# api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CS2KZ API",
    "description": "Source Code available on [GitHub](https://github.com/KZGlobalTeam/cs2kz-api).",
    "contact": {
      "name": "AlphaKeks",
      "email": "alphakeks@dawn.sh"
    },
    "license": {
      "name": "Licensed under the GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/servers": {
      "get": {
        "tags": [
          "Servers"
        ],
        "summary": "Fetch servers.",
        "operationId": "get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Filter by name.",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "host",
            "in": "query",
            "description": "Filter by host.\n\nThis can either be a domain name, or an IP address.",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "owned_by",
            "in": "query",
            "description": "Filter by server owner.",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/PlayerIdentifier"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "created_after",
            "in": "query",
            "description": "Only include servers approved after this date.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "created_before",
            "in": "query",
            "description": "Only include servers approved before this date.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "total",
                    "results"
                  ],
                  "properties": {
                    "total": {
                      "type": "integer",
                      "format": "uint64",
                      "minimum": 0
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "description": "A KZ server.",
                        "required": [
                          "id",
                          "name",
                          "host",
                          "port",
                          "owner",
                          "created_on"
                        ],
                        "properties": {
                          "id": {
                            "$ref": "#/components/schemas/ServerID"
                          },
                          "name": {
                            "type": "string",
                            "description": "The server's name."
                          },
                          "host": {
                            "$ref": "#/components/schemas/Host"
                          },
                          "port": {
                            "type": "integer",
                            "format": "uint16",
                            "description": "The server's port.",
                            "minimum": 0
                          },
                          "owner": {
                            "$ref": "#/components/schemas/Player"
                          },
                          "created_on": {
                            "type": "string",
                            "format": "date-time",
                            "description": "When this server was approved."
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Servers"
        ],
        "summary": "Create a new server.",
        "operationId": "post",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewServer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Response body for creating a new server.",
                  "required": [
                    "server_id",
                    "refresh_key"
                  ],
                  "properties": {
                    "server_id": {
                      "$ref": "#/components/schemas/ServerID"
                    },
                    "refresh_key": {
                      "type": "string",
                      "format": "uuid",
                      "description": "The server's API key."
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        },
        "security": [
          {
            "Browser Session": [
              "servers"
            ]
          }
        ]
      }
    },
    "/servers/{server}": {
      "get": {
        "tags": [
          "Servers"
        ],
        "summary": "Fetch a server by its name or ID.",
        "operationId": "get",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "A KZ server.",
                  "required": [
                    "id",
                    "name",
                    "host",
                    "port",
                    "owner",
                    "created_on"
                  ],
                  "properties": {
                    "id": {
                      "$ref": "#/components/schemas/ServerID"
                    },
                    "name": {
                      "type": "string",
                      "description": "The server's name."
                    },
                    "host": {
                      "$ref": "#/components/schemas/Host"
                    },
                    "port": {
                      "type": "integer",
                      "format": "uint16",
                      "description": "The server's port.",
                      "minimum": 0
                    },
                    "owner": {
                      "$ref": "#/components/schemas/Player"
                    },
                    "created_on": {
                      "type": "string",
                      "format": "date-time",
                      "description": "When this server was approved."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "Servers"
        ],
        "summary": "Update an existing server.",
        "operationId": "patch",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        },
        "security": [
          {
            "Browser Session": [
              "servers"
            ]
          }
        ]
      }
    },
...
    "/admins/{steam_id}": {
      "get": {
        "tags": [
          "Admins"
        ],
        "summary": "Fetch a specific admin by their SteamID.",
        "operationId": "get",
        "parameters": [
          {
            "name": "steam_id",
            "in": "path",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "string",
                  "title": "Steam ID",
                  "example": "STEAM_1:1:161178172"
                },
                {
                  "type": "string",
                  "title": "Steam ID3",
                  "example": "U:1:322356345"
                },
                {
                  "type": "integer",
                  "title": "Steam ID32",
                  "example": 322356345
                },
                {
                  "type": "integer",
                  "title": "Steam ID64",
                  "example": 76561198282622073
                }
              ],
              "description": "a player's SteamID",
              "example": "STEAM_1:1:161178172"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "A KZ admin.",
                  "required": [
                    "name",
                    "steam_id",
                    "permissions"
                  ],
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "The admin's name."
                    },
                    "steam_id": {
                      "$ref": "#/components/schemas/SteamID"
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "The admin's permissions.",
                      "example": [
                        "bans",
                        "servers"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Admins"
        ],
        "summary": "Create/Update an admin.",
        "description": "This endpoint is idempotent!",
        "operationId": "put",
        "parameters": [
          {
            "name": "steam_id",
            "in": "path",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "string",
                  "title": "Steam ID",
                  "example": "STEAM_1:1:161178172"
                },
                {
                  "type": "string",
                  "title": "Steam ID3",
                  "example": "U:1:322356345"
                },
                {
                  "type": "integer",
                  "title": "Steam ID32",
                  "example": 322356345
                },
                {
                  "type": "integer",
                  "title": "Steam ID64",
                  "example": 76561198282622073
                }
              ],
              "description": "a player's SteamID",
              "example": "STEAM_1:1:161178172"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        },
        "security": [
          {
            "Browser Session": [
              "admins"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminUpdate": {
        "type": "object",
        "description": "Request payload for updating an admin.",
        "required": [
          "permissions"
        ],
        "properties": {
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "New set of permissions for the admin.",
            "example": [
              "bans",
              "servers"
            ]
          }
        }
      },
...
      "Host": {
        "oneOf": [
          {
            "type": "string",
            "title": "IPv4",
            "example": "127.0.0.1"
          },
          {
            "type": "string",
            "title": "IPv6",
            "example": "::1"
          },
          {
            "type": "string",
            "title": "Domain",
            "format": "uri",
            "example": "example.com"
          }
        ],
        "description": "A server's host address"
      },
...
      "Limit": {
        "type": "number",
        "description": "limits the amount of returned values",
        "minimum": 0
      },
...
      "NewServer": {
        "type": "object",
        "description": "Request payload for creating a new server.",
        "required": [
          "name",
          "host",
          "port",
          "owned_by"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The server's name."
          },
          "host": {
            "type": "string",
            "description": "The server's host.\n\nThis can either be a domain name, or an IP address."
          },
          "port": {
            "type": "integer",
            "format": "uint16",
            "description": "The server's port.",
            "minimum": 0
          },
          "owned_by": {
            "$ref": "#/components/schemas/SteamID"
          }
        }
      },
...
      "Offset": {
        "type": "number",
        "description": "used for pagination",
        "default": 0,
        "maximum": 1.7976931348623157e308,
        "minimum": -1.7976931348623157e308
      },
      "Player": {
        "type": "object",
        "description": "Basic information about a KZ player.",
        "required": [
          "name",
          "steam_id"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The player's name."
          },
          "steam_id": {
            "$ref": "#/components/schemas/SteamID"
          }
        }
      },
      "PlayerIdentifier": {
        "oneOf": [
          {
            "oneOf": [
              {
                "type": "string",
                "title": "Steam ID",
                "example": "STEAM_1:1:161178172"
              },
              {
                "type": "string",
                "title": "Steam ID3",
                "example": "U:1:322356345"
              },
              {
                "type": "integer",
                "title": "Steam ID32",
                "example": 322356345
              },
              {
                "type": "integer",
                "title": "Steam ID64",
                "example": 76561198282622073
              }
            ],
            "description": "a player's SteamID",
            "example": "STEAM_1:1:161178172"
          },
          {
            "type": "string",
            "title": "Name"
          }
        ],
        "description": "A SteamID or name"
      },
...
      "ServerID": {
        "type": "integer",
        "format": "uint16",
        "minimum": 0
      },
...
      "ServerUpdate": {
        "type": "object",
        "description": "Request payload for updating a server.",
        "properties": {
          "name": {
            "type": "string",
            "description": "A new name.",
            "nullable": true
          },
          "host": {
            "type": "string",
            "description": "A new host.",
            "nullable": true
          },
          "port": {
            "type": "integer",
            "format": "uint16",
            "description": "A new port.",
            "nullable": true,
            "minimum": 0
          },
          "owned_by": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SteamID"
              }
            ],
            "nullable": true
          }
        }
      },
...
      "SteamID": {
        "oneOf": [
          {
            "type": "string",
            "title": "Steam ID",
            "example": "STEAM_1:1:161178172"
          },
          {
            "type": "string",
            "title": "Steam ID3",
            "example": "U:1:322356345"
          },
          {
            "type": "integer",
            "title": "Steam ID32",
            "example": 322356345
          },
          {
            "type": "integer",
            "title": "Steam ID64",
            "example": 76561198282622073
          }
        ],
        "description": "a player's SteamID",
        "example": "STEAM_1:1:161178172"
      },
...
    },
    "securitySchemes": {
      "API Key": {
        "type": "http",
        "scheme": "bearer"
      },
      "Browser Session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kz-auth"
      },
      "CS2 Server": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}